backend/requirements.txt
# --- Web Framework ---
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn>=21.2.0
python-multipart==0.0.6
fastapi-cors==0.0.6

# --- AI ---
google-generativeai==0.3.2  # Gemini SDK
httpx>=0.28.1               # Async client for Groq, Cohere, and GitHub

# --- Utilities ---
pydantic>=2.0.0
aiofiles>=23.2.1
python-dotenv==1.0.0
rich>=13.7.0

# --- Code Parsing (Tree-sitter for accurate AST analysis) ---
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-java>=0.21.0
tree-sitter-cpp>=0.21.0
tree-sitter-rust>=0.21.0

